requirements.txt
accelerate==0.25.0
bitsandbytes==0.41.2.post2
black==23.11.0
certifi==2023.11.17
charset-normalizer==3.3.2
chumpy==0.70
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
diffusers==0.24.0
filelock==3.13.1
fonttools==4.46.0
freetype-py==2.4.0
fsspec==2023.12.0
fvcore==0.1.5.post20221221
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.0
importlib-metadata==7.0.0
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.23.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
open3d-python==0.3.0.0
opencv-python==4.8.1.78
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
portalocker==2.8.2
psutil==5.9.6
pyglet==2.0.10
PyOpenGL==3.1.0
pyparsing==3.1.1
pyrender==0.1.45
python-dateutil==2.8.2
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@2f11ddc5ee7d6bd56f2fb6744a16776fab6536f7
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
shapely==2.0.2
six==1.16.0
smplx==0.1.28
sympy==1.12
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
trimesh==4.0.5
triton==2.1.0
typing_extensions==4.8.0
urllib3==2.1.0
uvtextureconverter==1.2.0
yacs==0.1.8
zipp==3.17.0
